// src/enums/HttpResponseCodes.ts
export enum HttpResponseCode {
        SUCCESS = 200,
        RESOURCES_CREATED = 201,
        NO_CONTENT = 204,
        BAD_REQUEST = 400,
        UNAUTHORIZED = 401,
        FORBIDDEN = 403,
        UNPROCESSABLE_ENTITY = 422,
        INTERNAL_SERVER_ERROR = 500,
        NOT_IMPLEMENTED = 501,
        CONFLICT = 409,
        NOT_FOUND = 404,
        EXCEPTION_FAILED = 417,
}